Create project.v
An ultrasonic distance measurement project in Verilog involves using an ultrasonic sensor to measure the distance to an object by calculating the time taken for sound waves to travel to the object and back. The Verilog module generates a trigger pulse for the sensor, captures the echo pulse duration using a timer, and converts the time into distance based on the speed of sound. The result is displayed  on seven segment which is inbuilt on FPGA.